Separated out test-driver and BuggyBlockingQueue from SynchronizedQueue class file, and created unit-test & proper package name usage
